requirements.txt
# Attendance Management System â€” Dependencies
# Install with: pip install -r requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

ultralytics==8.4.22
deepface==0.0.99
opencv-python==4.8.1.78
tensorflow_cpu==2.21.0
# tf-keras>=2.13.0        # Uncomment if TF 2.16+ gives keras errors
pandas==3.0.1
numpy==1.26.4
## The following requirements were added by pip freeze:
absl-py==2.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
astunparse==1.6.3
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Cython==3.2.4
easydict==1.13
faiss-cpu==1.13.2
filelock==3.25.2
fire==0.7.1
Flask==3.1.3
flask-cors==6.0.2
flatbuffers==25.12.19
fonttools==4.62.1
fsspec==2026.2.0
gast==0.7.0
gdown==5.2.1
google-pasta==0.2.0
grpcio==1.78.0
gunicorn==25.1.0
h5py==3.14.0
idna==3.11
ImageIO==2.37.3
insightface==0.7.3
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
keras==3.13.2
kiwisolver==1.5.0
lazy-loader==0.5
libclang==18.1.1
lightdsa==0.0.3
lightecc==0.0.4
lightphe==0.0.20
lz4==4.4.5
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
mtcnn==1.0.0
mysql-connector-python==9.6.0
namex==0.1.0
networkx==3.6.1
omegaconf==2.3.0
onnx==1.20.1
onnxruntime==1.24.3
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pillow==12.1.1
polars==1.39.0
polars-runtime-32==1.39.0
prettytable==3.17.0
protobuf==7.34.0
psutil==7.2.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.32.5
retina-face==0.0.17
rich==14.3.3
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==82.0.1
simsimd==6.5.16
six==1.17.0
soupsieve==2.8.3
stringzilla==4.6.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
tf_keras==2.20.1
threadpoolctl==3.6.0
tifffile==2026.3.3
torch==2.10.0+cpu
torchvision==0.25.0+cpu
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ultralytics-thop==2.0.18
urllib3==2.6.3
wcwidth==0.6.0
Werkzeug==3.1.6
wheel==0.46.3
wrapt==2.1.2